Resolves #1887: Splits SyntheticRecordPlannerTest into multiple classes
