+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/query/plan/synthetic/AbstractSyntheticRecordPlannerTest.java
@@ -20,6 +20,8 @@
 
 package com.apple.foundationdb.record.query.plan.synthetic;
 
+import com.apple.foundationdb.annotation.API;
+import com.apple.foundationdb.record.metadata.JoinedRecordType;
 import com.apple.foundationdb.record.provider.foundationdb.FDBDatabase;
 import com.apple.foundationdb.record.provider.foundationdb.FDBStoreTimer;
 import com.apple.foundationdb.record.RecordMetaData;
@@ -27,16 +29,31 @@ import com.apple.foundationdb.record.RecordMetaDataBuilder;
 import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStore;
 import com.apple.foundationdb.record.TestRecordsJoinIndexProto;
 import com.apple.foundationdb.record.provider.foundationdb.FDBRecordContext;
+import com.apple.foundationdb.record.query.plan.PlannableIndexTypes;
+import com.apple.foundationdb.record.query.plan.QueryPlanner;
+import com.apple.foundationdb.record.query.plan.RecordQueryPlanner;
 import com.apple.foundationdb.record.test.FDBDatabaseExtension;
 import com.apple.foundationdb.record.test.TestKeySpacePathManagerExtension;
 import com.apple.foundationdb.record.test.TestKeySpace;
 import com.apple.foundationdb.record.provider.foundationdb.keyspace.KeySpacePath;
+import com.apple.test.Tags;
+import org.hamcrest.Matcher;
+import org.hamcrest.Matchers;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.extension.RegisterExtension;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import java.util.Map;
+
+import static org.hamcrest.MatcherAssert.assertThat;
 
 /**
  * Base class for {@link SyntheticRecordPlanner} tests.
  */
+@Tag(Tags.RequiresFDB)
+@API(API.Status.EXPERIMENTAL)
 public abstract class AbstractSyntheticRecordPlannerTest {
     @RegisterExtension
     final FDBDatabaseExtension dbExtension = new FDBDatabaseExtension();
@@ -50,6 +67,24 @@ public abstract class AbstractSyntheticRecordPlannerTest {
 
     protected FDBRecordContext openContext() {
         return fdb.openContext(null, timer);
+    }
+
+    protected QueryPlanner setupPlanner(@Nonnull FDBRecordStore recordStore, @Nullable PlannableIndexTypes indexTypes) {
+        if (indexTypes == null) {
+            indexTypes = PlannableIndexTypes.DEFAULT;
+        }
+        return new RecordQueryPlanner(recordStore.getRecordMetaData(), recordStore.getRecordStoreState(), indexTypes, recordStore.getTimer());
+    }
+
+    protected static void assertConstituentPlansMatch(SyntheticRecordPlanner planner, JoinedRecordType joinedRecordType,
+                                                    Map<String, Matcher<? super SyntheticRecordFromStoredRecordPlan>> constituentMatchers) {
+        for (JoinedRecordType.JoinConstituent constituent : joinedRecordType.getConstituents()) {
+            assertThat("constituent matchers missing matcher for constituent " + constituent.getName(),
+                    constituentMatchers, Matchers.hasKey(constituent.getName()));
+            Matcher<? super SyntheticRecordFromStoredRecordPlan> matcher = constituentMatchers.get(constituent.getName());
+            final SyntheticRecordFromStoredRecordPlan plan = planner.forJoinConstituent(joinedRecordType, constituent);
+            assertThat(plan, matcher);
+        }
     }
 
     @BeforeEach
